+++ packages/dialtone-vue/components/input/input.vue
@@ -166,7 +166,7 @@ export default {
     /**
      * Type of the input.
      * When `textarea` a `<textarea>` element will be rendered instead of an `<input>` element.
-     * @values text, password, email, number, textarea, date, time, file, tel, search
+     * @values text, password, email, number, textarea, date, time, file, tel, search, color
      * @default 'text'
      */
     type: {
@@ -373,6 +373,7 @@ export default {
       isInvalid: false,
       defaultLength: 0,
       hasSlotContent,
+      isComposing: false,
     };
   },
 
@@ -408,7 +409,16 @@ export default {
 
     inputListeners () {
       return {
+        compositionstart: () => {
+          this.isComposing = true;
+        },
+
+        compositionend: () => {
+          this.isComposing = false;
+        },
+
         input: async event => {
+          if (this.isComposing) return;
           let val = event.target.value;
           if (this.type === INPUT_TYPES.FILE) {
             const files = Array.from(event.target.files);
@@ -534,7 +544,8 @@ export default {
         }
 
         // Set textarea value programmatically to avoid attribute binding
-        if (this.isTextarea && this.$refs.input && this.$refs.input.value !== newValue) {
+        // Skip during IME composition to avoid interrupting in-progress input
+        if (this.isTextarea && this.$refs.input && this.$refs.input.value !== newValue && !this.isComposing) {
           this.$refs.input.value = newValue;
         }
       },