+++ packages/dialtone-vue/recipes/conversation_view/editor/editor.vue
@@ -18,111 +18,213 @@
         direction="row"
         gap="300"
       >
-        <dt-tooltip
-          v-for="button in buttonGroup.buttonGroup"
-          :key="getButtonKey(buttonGroup.key, button.selector)"
-          :message="button.tooltipMessage"
-          placement="top"
-        >
-          <template #anchor>
-            <dt-button
-              :ref="getButtonRef(buttonGroup.key, button.selector)"
-              :active="$refs.richTextEditor?.editor?.isActive(button.selector)"
-              :aria-label="button.tooltipMessage"
-              :data-qa="button.dataQA"
-              :tabindex="canFocus(getButtonRef(buttonGroup.key, button.selector)) ? 0 : -1"
-              importance="clear"
-              kind="muted"
-              size="xs"
-              @click="button.onClick()"
-              @keydown.right.stop="shiftActionBarFocusRight"
-              @keydown.left.stop="shiftActionBarFocusLeft"
-            >
-              <template #icon>
-                <component
-                  :is="button.icon"
-                  size="200"
-                />
-              </template>
-              {{ button?.label }}
-            </dt-button>
-          </template>
-        </dt-tooltip>
-        <div class="d-recipe-editor__button-group-divider" />
-      </dt-stack>
-      <dt-stack
-        v-if="variableButton.showBtn"
-        direction="row"
-        gap="300"
-      >
-        <dt-popover
-          padding="small"
-          navigation-type="arrow-keys"
-          :modal="false"
-          placement="bottom-start"
-        >
-          <template #anchor="{ attrs }">
-            <dt-tooltip
-              :message="variableButton.tooltipMessage"
-              placement="top"
-            >
-              <template #anchor>
-                <dt-button
-                  v-bind="attrs"
-                  kind="muted"
-                  size="xs"
-                  importance="clear"
-                  :aria-label="variableButton.tooltipMessage"
-                  :data-qa="variableButton.dataQA"
-                  label-class="d-jc-flex-start"
-                >
-                  <template #icon>
-                    <component
-                      :is="variableButton.icon"
-                      size="200"
-                    />
-                  </template>
-                </dt-button>
-              </template>
-            </dt-tooltip>
-          </template>
-          <template #content="{ close }">
-            <dt-input
-              v-model="variableSearchValue"
-              root-class="d-p8 d-pb4 d-w264"
-              type="search"
-              :placeholder="i18n.$t('DIALTONE_EDITOR_VARIABLE_POPOVER_SEARCH_PLACEHOLDER')"
-              size="md"
-              role="menuitem"
-            >
-              <template #leftIcon="{ iconSize }">
-                <dt-icon-search :size="iconSize" />
-              </template>
-            </dt-input>
-            <dt-list-item-group
-              v-for="(category, index) in filteredCategories"
-              :key="category.name"
-              :heading="category.name"
-              heading-class="d-headline--sm-compact d-p8"
-            >
+        <template v-for="button in buttonGroup.buttonGroup">
+          <!-- Font Style Popover -->
+          <editor-toolbar-popover-button
+            v-if="button.buttonType === 'popover' && button.selector === 'fontStyle'"
+            :key="getButtonKey(buttonGroup.key, button.selector)"
+            :ref="getButtonRef(buttonGroup.key, button.selector)"
+            :is-active="$refs.richTextEditor?.editor?.isActive(button.selector)"
+            :tooltip-message="button.tooltipMessage"
+            :data-qa="button.dataQA"
+            :popover-data-qa="'dt-recipe-editor-font-style-input-popover'"
+            :tabindex="canFocus(getButtonRef(buttonGroup.key, button.selector)) ? 0 : -1"
+            :icon="button.icon"
+            @shift-focus-right="shiftActionBarFocusRight"
+            @shift-focus-left="shiftActionBarFocusLeft"
+          >
+            <template #content="{ close }">
+              <dt-combobox
+                label=""
+                :label-visible="false"
+                :show-list="true"
+                :click-on-select="true"
+                @escape="close()"
+              >
+                <template #input="{ inputProps }">
+                  <dt-input
+                    v-bind="inputProps"
+                    v-model="fontStyleSearch"
+                    root-class="d-p8 d-pb4 d-w216"
+                    type="search"
+                    :placeholder="i18n.$t('DIALTONE_EDITOR_FONT_STYLE_SEARCH_PLACEHOLDER')"
+                    size="sm"
+                    role="menuitem"
+                  >
+                    <template #leftIcon="{ iconSize }">
+                      <dt-icon-search :size="iconSize" />
+                    </template>
+                  </dt-input>
+                </template>
+                <template #list="{ listProps }">
+                  <ul
+                    v-bind="listProps"
+                  >
+                    <dt-list-item
+                      v-for="fontStyle in filteredFontStyles"
+                      :key="fontStyle.name"
+                      :selected="isCurrentFontFamily(fontStyle.value)"
+                      :style="{ fontFamily: fontStyle.value || 'inherit' }"
+                      role="option"
+                      navigation-type="arrow-keys"
+                      @click="
+                        close();
+                        onFontStyleSelect(fontStyle.value)
+                      "
+                    >
+                      {{ fontStyle.name }}
+                    </dt-list-item>
+                  </ul>
+                </template>
+              </dt-combobox>
+            </template>
+          </editor-toolbar-popover-button>
+
+          <!-- Font Size Dropdown -->
+          <editor-toolbar-dropdown-button
+            v-else-if="button.buttonType === 'popover' && button.selector === 'fontSize'"
+            :key="getButtonKey(buttonGroup.key, button.selector)"
+            :ref="getButtonRef(buttonGroup.key, button.selector)"
+            :is-active="$refs.richTextEditor?.editor?.isActive(button.selector)"
+            :tooltip-message="button.tooltipMessage"
+            :data-qa="button.dataQA"
+            :dropdown-data-qa="'dt-recipe-editor-font-size-input-popover'"
+            :tabindex="canFocus(getButtonRef(buttonGroup.key, button.selector)) ? 0 : -1"
+            :icon="button.icon"
+            @shift-focus-right="shiftActionBarFocusRight"
+            @shift-focus-left="shiftActionBarFocusLeft"
+          >
+            <template #list="{ close }">
               <dt-list-item
-                v-for="item in getFilteredItemsForCategory(category)"
-                :key="category.name + item.name"
+                v-for="fontSize in fontSizes"
+                :key="fontSize.name"
+                :selected="isCurrentFontSize(fontSize.value)"
                 role="menuitem"
                 navigation-type="arrow-keys"
                 @click="
-                  insertVariable(category.name, item);
                   close();
+                  onFontSizeSelect(fontSize.value, $event)
                 "
               >
-                {{ item.name }}
+                <span :style="{ fontSize: fontSize.value }">{{ fontSize.name }}</span>
               </dt-list-item>
-              <dt-dropdown-separator
-                v-if="index < filteredCategories.length - 1"
+            </template>
+          </editor-toolbar-dropdown-button>
+
+          <!-- Font Color Button -->
+          <dt-button
+            v-else-if="button.buttonType === 'custom' && button.selector === 'fontColor'"
+            :key="getButtonKey(buttonGroup.key, button.selector)"
+            :ref="getButtonRef(buttonGroup.key, button.selector)"
+            v-dt-tooltip="{ message: button.tooltipMessage, placement: 'top' }"
+            kind="muted"
+            importance="clear"
+            size="xs"
+            :active="$refs.richTextEditor?.editor?.isActive(button.selector)"
+            :tabindex="canFocus(getButtonRef(buttonGroup.key, button.selector)) ? 0 : -1"
+            :aria-label="button.tooltipMessage"
+            :data-qa="button.dataQA"
+            @keydown.right.stop="shiftActionBarFocusRight"
+            @keydown.left.stop="shiftActionBarFocusLeft"
+            @click="button.onClick()"
+          >
+            <template #default>
+              <dt-input
+                :value="currentFontColor"
+                root-class="d-w12 d-h12 d-my1"
+                input-class="colorPickerInput d-w100p d-h100p d-p0 d-bar0 d-c-pointer"
+                input-wrapper-class="d-w12 d-h12 d-bar2 d-ba-none"
+                size="sm"
+                type="color"
+                @input="onColorPickerInput"
+                @click.stop
               />
-            </dt-list-item-group>
-          </template>
-        </dt-popover>
+            </template>
+          </dt-button>
+
+          <!-- Variable Popover -->
+          <editor-toolbar-popover-button
+            v-else-if="button.buttonType === 'popover' && button.selector === 'variable'"
+            :key="getButtonKey(buttonGroup.key, button.selector)"
+            :ref="getButtonRef(buttonGroup.key, button.selector)"
+            :is-active="false"
+            :tooltip-message="button.tooltipMessage"
+            :data-qa="button.dataQA"
+            :popover-data-qa="'dt-recipe-editor-variable-popover'"
+            :tabindex="canFocus(getButtonRef(buttonGroup.key, button.selector)) ? 0 : -1"
+            :icon="button.icon"
+            @shift-focus-right="shiftActionBarFocusRight"
+            @shift-focus-left="shiftActionBarFocusLeft"
+          >
+            <template #content="{ close }">
+              <dt-combobox
+                label=""
+                :label-visible="false"
+                :show-list="true"
+                :click-on-select="true"
+                @escape="close()"
+              >
+                <template #input="{ inputProps }">
+                  <dt-input
+                    v-bind="inputProps"
+                    v-model="variableSearchValue"
+                    root-class="d-p8 d-pb4 d-w264"
+                    type="search"
+                    :placeholder="i18n.$t('DIALTONE_EDITOR_VARIABLE_POPOVER_SEARCH_PLACEHOLDER')"
+                    size="md"
+                    role="menuitem"
+                  >
+                    <template #leftIcon="{ iconSize }">
+                      <dt-icon-search :size="iconSize" />
+                    </template>
+                  </dt-input>
+                </template>
+                <template #list="{ listProps }">
+                  <div v-bind="listProps">
+                    <dt-list-item-group
+                      v-for="(category, index) in filteredCategories"
+                      :key="category.name"
+                      :heading="category.name"
+                      heading-class="d-headline--sm-compact d-p8"
+                    >
+                      <dt-list-item
+                        v-for="item in getFilteredItemsForCategory(category)"
+                        :key="category.name + item.name"
+                        role="option"
+                        navigation-type="arrow-keys"
+                        @click="
+                          insertVariable(category.name, item);
+                          close();
+                        "
+                      >
+                        {{ item.name }}
+                      </dt-list-item>
+                      <dt-dropdown-separator
+                        v-if="index < filteredCategories.length - 1"
+                      />
+                    </dt-list-item-group>
+                  </div>
+                </template>
+              </dt-combobox>
+            </template>
+          </editor-toolbar-popover-button>
+
+          <!-- Regular Toolbar Button -->
+          <editor-toolbar-button
+            v-else-if="button.buttonType === 'button'"
+            :key="getButtonKey(buttonGroup.key, button.selector)"
+            :ref="getButtonRef(buttonGroup.key, button.selector)"
+            :is-active="$refs.richTextEditor?.editor?.isActive(button.selector)"
+            :tooltip-message="button.tooltipMessage"
+            :data-qa="button.dataQA"
+            :tabindex="canFocus(getButtonRef(buttonGroup.key, button.selector)) ? 0 : -1"
+            :icon="button.icon"
+            :label="button.label"
+            :on-click="button.onClick"
+            @shift-focus-right="shiftActionBarFocusRight"
+            @shift-focus-left="shiftActionBarFocusLeft"
+          />
+        </template>
         <div class="d-recipe-editor__button-group-divider" />
       </dt-stack>
       <dt-stack
@@ -242,6 +344,7 @@
         :allow-inline-images="true"
         :allow-line-breaks="true"
         :allow-variable="true"
+        :allow-font-size="showFontSizeButton"
         :variable-items="flattenedVariableItems"
         :hide-link-bubble-menu="true"
         :auto-focus="autoFocus"
@@ -260,13 +363,13 @@
         @blur="onBlur"
         @focus="onFocus"
         @input="onInput($event)"
+        @selected="onSelected"
       />
     </div>
   </div>
 </template>
 
 <script>
-/* eslint-disable max-lines */
 import {
   DtRichTextEditor,
   RICH_TEXT_EDITOR_OUTPUT_FORMATS,
@@ -275,6 +378,7 @@ import {
 import {
   EDITOR_SUPPORTED_LINK_PROTOCOLS,
   EDITOR_DEFAULT_LINK_PREFIX,
+  EDITOR_DEFAULT_FONT_COLOR,
 } from './editor_constants.js';
 import { removeClassStyleAttrs, addClassStyleAttrs } from '@/common/utils';
 import { DtButton } from '@/components/button';
@@ -282,9 +386,14 @@ import { DtPopover } from '@/components/popover';
 import { DtStack } from '@/components/stack';
 import { DtInput } from '@/components/input';
 import { DtTooltip } from '@/components/tooltip';
-import {DtListItem} from '@/components/list_item/index.js';
-import {DtDropdownSeparator} from '@/components/dropdown/index.js';
-import {DtListItemGroup} from '@/components/list_item_group/index.js';
+import { DtListItem } from '@/components/list_item';
+import { DtCombobox } from '@/components/combobox';
+
+import { DtDropdownSeparator } from '@/components/dropdown/index.js';
+import { DtListItemGroup } from '@/components/list_item_group/index.js';
+import EditorToolbarButton from './EditorToolbarButton.vue';
+import EditorToolbarDropdownButton from './EditorToolbarDropdownButton.vue';
+import EditorToolbarPopoverButton from './EditorToolbarPopoverButton.vue';
 import {
   DtIconAlignCenter,
   DtIconAlignJustify,
@@ -301,8 +410,11 @@ import {
   DtIconQuote,
   DtIconStrikethrough,
   DtIconUnderline,
+  DtIconType,
   DtIconBraces,
   DtIconSearch,
+  DtIconChevronDown,
+  DtIconFontSize,
 } from '@dialpad/dialtone-icons/vue3';
 import { DialtoneLocalization } from '@/localization';
 
@@ -320,6 +432,10 @@ export default {
     DtStack,
     DtInput,
     DtTooltip,
+    DtCombobox,
+    EditorToolbarButton,
+    EditorToolbarDropdownButton,
+    EditorToolbarPopoverButton,
     DtIconQuickReply,
     DtIconBold,
     DtIconItalic,
@@ -335,8 +451,11 @@ export default {
     DtIconCodeBlock,
     DtIconLink2,
     DtIconImage,
-    DtIconBraces,
     DtIconSearch,
+    DtIconType,
+    DtIconBraces,
+    DtIconChevronDown,
+    DtIconFontSize,
   },
 
   mixins: [],
@@ -566,6 +685,54 @@ export default {
     },
 
     /**
+     * Show font style button.
+     */
+    showFontStyleButton: {
+      type: Boolean,
+      default: false,
+    },
+
+    /**
+     * Show font size button.
+     */
+    showFontSizeButton: {
+      type: Boolean,
+      default: false,
+    },
+
+    /**
+     * Show font color button.
+     */
+    showFontColorButton: {
+      type: Boolean,
+      default: false,
+    },
+
+    /**
+     * Available font styles for the font style dropdown.
+     */
+    fontStyles: {
+      type: Array,
+      default: () => [
+        { name: 'Arial', value: null }, // arial is the default font
+        { name: 'Georgia', value: 'Georgia' },
+        { name: 'Helvetica', value: 'Helvetica' },
+        { name: 'Verdana', value: 'Verdana'},
+        { name: 'Times New Roman', value: 'Times New Roman' },
+      ],
+    },
+
+    fontSizes : {
+      type: Array,
+      default: () => [
+        { name: 'Small', value: '12px'},
+        { name: 'Normal', value: '15px'},
+        { name: 'Large', value: '24px'},
+        { name: 'Huge', value: '36px'},
+      ],
+    },
+
+    /**
      * Use div tags instead of paragraph tags to show text
      */
     useDivTags: {
@@ -649,7 +816,9 @@ export default {
         class: 'd-recipe-editor__link',
       },
 
+      currentFontColor: undefined,
       showLinkInput: false,
+      fontStyleSearch: '',
       linkInput: '',
       currentButtonRefIndex: 0,
       variableSearchValue: '',
@@ -715,6 +884,7 @@ export default {
       return [
         {
           showBtn: this.showQuickRepliesButton,
+          buttonType: 'button',
           label: this.i18n.$t('DIALTONE_EDITOR_QUICK_REPLY_BUTTON_LABEL'),
           selector: 'quickReplies',
           icon: DtIconQuickReply,
@@ -728,7 +898,32 @@ export default {
     textFormatButtons () {
       return [
         {
+          showBtn: this.showFontStyleButton,
+          buttonType: 'popover',
+          selector: 'fontStyle',
+          icon: DtIconType,
+          dataQA: 'dt-recipe-editor-font-style-btn',
+          tooltipMessage: this.i18n.$t('DIALTONE_EDITOR_FONT_STYLE_BUTTON_LABEL'),
+        },
+        {
+          showBtn: this.showFontSizeButton,
+          buttonType: 'popover',
+          selector: 'fontSize',
+          icon: DtIconFontSize,
+          dataQA: 'dt-recipe-editor-font-size-btn',
+          tooltipMessage: this.i18n.$t('DIALTONE_EDITOR_FONT_SIZE_BUTTON_LABEL'),
+        },
+        {
+          showBtn: this.showFontColorButton,
+          buttonType: 'custom',
+          selector: 'fontColor',
+          dataQA: 'dt-recipe-editor-font-color-btn',
+          tooltipMessage: this.i18n.$t('DIALTONE_EDITOR_FONT_COLOR_BUTTON_LABEL'),
+          onClick: this.onColorPickerButtonClick,
+        },
+        {
           showBtn: this.showBoldButton,
+          buttonType: 'button',
           selector: 'bold',
           icon: DtIconBold,
           dataQA: 'dt-recipe-editor-bold-btn',
@@ -737,6 +932,7 @@ export default {
         },
         {
           showBtn: this.showItalicsButton,
+          buttonType: 'button',
           selector: 'italic',
           icon: DtIconItalic,
           dataQA: 'dt-recipe-editor-italics-btn',
@@ -745,6 +941,7 @@ export default {
         },
         {
           showBtn: this.showUnderlineButton,
+          buttonType: 'button',
           selector: 'underline',
           icon: DtIconUnderline,
           dataQA: 'dt-recipe-editor-underline-btn',
@@ -753,6 +950,7 @@ export default {
         },
         {
           showBtn: this.showStrikeButton,
+          buttonType: 'button',
           selector: 'strike',
           icon: DtIconStrikethrough,
           dataQA: 'dt-recipe-editor-strike-btn',
@@ -766,6 +964,7 @@ export default {
       return [
         {
           showBtn: this.showAlignLeftButton,
+          buttonType: 'button',
           selector: { textAlign: 'left' },
           icon: DtIconAlignLeft,
           dataQA: 'dt-recipe-editor-align-left-btn',
@@ -774,6 +973,7 @@ export default {
         },
         {
           showBtn: this.showAlignCenterButton,
+          buttonType: 'button',
           selector: { textAlign: 'center' },
           icon: DtIconAlignCenter,
           dataQA: 'dt-recipe-editor-align-center-btn',
@@ -782,6 +982,7 @@ export default {
         },
         {
           showBtn: this.showAlignRightButton,
+          buttonType: 'button',
           selector: { textAlign: 'right' },
           icon: DtIconAlignRight,
           dataQA: 'dt-recipe-editor-align-right-btn',
@@ -790,6 +991,7 @@ export default {
         },
         {
           showBtn: this.showAlignJustifyButton,
+          buttonType: 'button',
           selector: { textAlign: 'justify' },
           icon: DtIconAlignJustify,
           dataQA: 'dt-recipe-editor-align-justify-btn',
@@ -803,6 +1005,7 @@ export default {
       return [
         {
           showBtn: this.showListItemsButton,
+          buttonType: 'button',
           selector: 'bulletList',
           icon: DtIconListBullet,
           dataQA: 'dt-recipe-editor-list-items-btn',
@@ -811,6 +1014,7 @@ export default {
         },
         {
           showBtn: this.showOrderedListButton,
+          buttonType: 'button',
           selector: 'orderedList',
           icon: DtIconListOrdered,
           dataQA: 'dt-recipe-editor-ordered-list-items-btn',
@@ -824,6 +1028,7 @@ export default {
       return [
         {
           showBtn: this.showQuoteButton,
+          buttonType: 'button',
           selector: 'blockquote',
           icon: DtIconQuote,
           dataQA: 'dt-recipe-editor-blockquote-btn',
@@ -832,6 +1037,7 @@ export default {
         },
         {
           showBtn: this.showCodeBlockButton,
+          buttonType: 'button',
           selector: 'codeBlock',
           icon: DtIconCodeBlock,
           dataQA: 'dt-recipe-editor-code-block-btn',
@@ -840,6 +1046,7 @@ export default {
         },
         {
           showBtn: this.showInlineImageButton,
+          buttonType: 'button',
           selector: 'image',
           icon: DtIconImage,
           dataQA: 'dt-recipe-editor-inline-image-btn',
@@ -847,12 +1054,21 @@ export default {
           // Handle getting image
           onClick: this.onInsertInlineImageClick,
         },
+        {
+          showBtn: this.showVariableButton,
+          buttonType: 'popover',
+          selector: 'variable',
+          icon: DtIconBraces,
+          dataQA: 'dt-recipe-editor-variable-btn',
+          tooltipMessage: this.i18n.$t('DIALTONE_EDITOR_VARIABLE_BUTTON_LABEL'),
+        },
       ].filter(button => button.showBtn);
     },
 
     linkButton () {
       return {
         showBtn: this.showAddLink.showAddLinkButton,
+        buttonType: 'custom',
         selector: 'link',
         icon: DtIconLink2,
         dataQA: 'dt-recipe-editor-add-link-btn',
@@ -860,17 +1076,6 @@ export default {
         onClick: this.openLinkInput,
       };
     },
-
-    variableButton() {
-      return {
-        showBtn: this.showVariableButton,
-        selector: 'variable',
-        icon: DtIconBraces,
-        dataQA: 'dt-recipe-editor-variable-btn',
-        tooltipMessage: this.i18n.$t('DIALTONE_EDITOR_VARIABLE_BUTTON_LABEL'),
-      }
-    },
-
 
     confirmSetLinkButtonLabels () {
       return this.i18n.$ta('DIALTONE_EDITOR_CONFIRM_SET_LINK_BUTTON');
@@ -888,11 +1093,29 @@ export default {
       return this.i18n.$ta('DIALTONE_EDITOR_ADD_LINK_BUTTON');
     },
 
+    filteredFontStyles () {
+      const searchValue = this.fontStyleSearch.toLowerCase();
+      return this.fontStyles.filter((item) =>
+        item.name.toLowerCase().includes(searchValue),
+      );
+    },
+
     filteredCategories() {
       return this.variableCategories.filter(
         (category) => this.getFilteredItemsForCategory(category).length,
       );
     },
+
+    colorPickerInput() {
+      return document.querySelector('.colorPickerInput');
+    },
+
+    actionBarBtn () {
+      const ref = this.$refs[this.orderedRefs[this.currentButtonRefIndex]][0]?.$refs?.buttonRef // get nested ref
+        || this.$refs[this.orderedRefs[this.currentButtonRefIndex]];
+      return Array.isArray(ref) ? ref[0] : ref;
+    },
+
   },
 
   watch: {
@@ -1066,6 +1289,10 @@ export default {
       this.$emit('update:modelValue', event);
     },
 
+    onSelected() {
+      this.updateFontColorInput();
+    },
+
     getButtonKey (key, selector) {
       return `${key}-${JSON.stringify(selector)}`;
     },
@@ -1092,14 +1319,41 @@ export default {
     },
 
     shiftButtonRefIndex (shiftAmount) {
-      const previousRef = this.$refs[this.orderedRefs[this.currentButtonRefIndex]];
-      const previousActionBarBtn = Array.isArray(previousRef) ? previousRef[0] : previousRef;
+      const previousActionBarBtn = this.actionBarBtn;
       const index = (this.currentButtonRefIndex + shiftAmount) % this.orderedRefs.length;
       this.currentButtonRefIndex = index >= 0 ? index : this.orderedRefs.length + index;
-      const currentRef = this.$refs[this.orderedRefs[this.currentButtonRefIndex]];
-      const currentActionBarBtn = Array.isArray(currentRef) ? currentRef[0] : currentRef;
+      const currentActionBarBtn = this.actionBarBtn;
+
       previousActionBarBtn.$el.blur();
       currentActionBarBtn.$el.focus();
+    },
+
+    onFontStyleSelect (fontFamily) {
+      if (fontFamily) {
+        this.$refs.richTextEditor?.editor?.chain().focus().setFontFamily(fontFamily).run();
+      } else {
+        this.$refs.richTextEditor?.editor?.chain().focus().unsetFontFamily().run();
+      }
+      this.$refs.richTextEditor?.editor?.commands.focus();
+    },
+
+    isCurrentFontFamily (fontFamily) {
+      if (!fontFamily) {
+        return !this.$refs.richTextEditor?.editor?.getAttributes('textStyle')?.fontFamily;
+      }
+      return this.$refs.richTextEditor?.editor?.isActive('textStyle', { fontFamily });
+    },
+
+    onFontSizeSelect (fontSize) {
+      this.$refs.richTextEditor?.editor?.chain().focus().setFontSize(fontSize).run();
+      this.$refs.richTextEditor?.editor?.commands.focus();
+    },
+
+    isCurrentFontSize (fontSize) {
+      if (!fontSize) {
+        return !this.$refs.richTextEditor?.editor?.getAttributes('textStyle')?.fontSize;
+      }
+      return this.$refs.richTextEditor?.editor?.isActive('textStyle', { fontSize });
     },
 
     getFilteredItemsForCategory(category) {
@@ -1110,6 +1364,19 @@ export default {
       return category.items.filter((item) =>
         item.name.toLowerCase().includes(searchValue),
       );
+    },
+
+    updateFontColorInput() {
+      this.currentFontColor = this.$refs.richTextEditor?.editor?.getAttributes('textStyle')?.color || EDITOR_DEFAULT_FONT_COLOR;
+    },
+
+    onColorPickerButtonClick () {
+      this.colorPickerInput?.click();
+    },
+
+    onColorPickerInput (fontColor) {
+      this.$refs.richTextEditor?.editor?.chain().focus().setColor(fontColor).run();
+      this.$refs.richTextEditor?.editor?.commands.focus();
     },
   },
 };
@@ -1153,5 +1420,11 @@ export default {
 .d-recipe-editor__content-image-resize .tiptap .ProseMirror-selectednode [data-resize-handle='bottom-right'] {
   cursor: nwse-resize;
   transform: translate(50%, 50%);
+}
+.colorPickerInput::-webkit-color-swatch-wrapper {
+  padding: 0 !important;
+}
+.colorPickerInput::-webkit-color-swatch {
+  border: none !important;
 }
 </style>